requirements.txt
fastapi 
fastapi-mcp 
uvicorn
fastmcp
feedparser
strands-agents
strands-agents-tools
langchain_groq